+++ src/main/dll/DIM/dll_0256_dimsnowhorn1.c
@@ -800,7 +800,7 @@ int DIMSnowHorn1_stateHandler0A(int obj, int state, f32 t)
     }
 
     ((ObjAnimSampleRootCurveObjectFirstFn)ObjAnim_SampleRootCurvePhase)(
-        (int)obj, ((DIMSnowHorn1State*)state)->baddie.animSpeedA,
+        obj, ((DIMSnowHorn1State*)state)->baddie.animSpeedA,
         &((DIMSnowHorn1State*)state)->baddie.moveSpeed);
     if ((*(int*)&((DIMSnowHorn1State*)state)->baddie.unk31C & 0x100) != 0)
     {
